Change the approach used to index source file contents to instead store a list of source code fragments (class, function decls, function calls etc.) that can be indexed individually.
These fragments are determined by parsing the source file using Roslyn and
then traversing the AST to find "interesting" nodes.

Indexing an entire source file proved difficult with regards to getting the
correct amount of content info to show (as the amount of context may change
depending on where the match is).

This way we can highlight the entire fragment (since they'll be quite small)
and we'll get all the required context needed.

Still need to figure out how to get the line numbers from Roslyn but this is
a reasonable first cut to start fleshing out the client support.